Share pre-read seams across runtime adapters
Claude and Codex both use the same pre-read decision and prompt context policies, but Claude previously imported Codex-named adapter modules directly. This separates the shared seams into neutral adapter modules while keeping the existing Codex compatibility export for CLI/tests.

Constraint: Keep the existing decideCodexPreRead seam mutable for current regression tests and runtime fallback coverage.

Rejected: Leave Claude importing codex-pre-read | preserves behavior but keeps misleading Codex-specific coupling in shared runtime code.

Rejected: Re-export decideCodexPreRead as an alias only | Node module namespace exports become read-only and break the existing monkey-patch regression tests.

Confidence: high

Scope-risk: narrow

Directive: Keep codex-pre-read.ts as the compatibility boundary unless tests and CLI callers are migrated together.

Tested: npm run lint

Tested: npm test (194/194)

Tested: npm run release:smoke
