+++ src/adapters/prompt-context.ts
@@ -1,0 +1,10 @@
+export {
+  hasFullReadEscapeHatch,
+  resolvePromptFileContext,
+} from "../core/context-policy";
+
+export type {
+  ContextBudget,
+  ContextMode,
+  PromptSpecificity,
+} from "../core/context-policy";
